requirements.txt
boto3==1.34.84
botocore==1.34.84
sagemaker==2.214.0
langchain==0.2.1
langgraph==0.0.55
streamlit==1.35.0
python-dotenv==1.0.1
joblib==1.4.2
xgboost==1.2.0